Filter recipes by 'types' of diets created
Currently only vegan and vegetarian are working due to the state of
recipe table.

There is scafolding however for the other diets pecestarian, keto etc.
to be added in with relatively little work, using abstract functions and
copying generic sql procedures.

Currently the filter is an OR statement in that if multiple filters are
selected e.g. ?vegan=true&vegetarian=true.
You will receieve both vegan and vegetarian recipes rather than
recipes that are both vegan AND vegetarian.

Duplication filtering is done via python after multiple sql queries
for ease of development. Talks to improve this by moving logic into
sql procedures can be for the future.