Grocery List Frontend Completed
Grocery List Frontend has been made but need to work on what happens when a section is empty: right now there is a section footer but hoping to render an empty item or create some whitespace so it doesn't look so weird.

TODO:

1. Need a button to go between the grocery list and the other pages in the fridge tab.

2. Hook up the grocery list backend to the front end. The page is currently using sample data but the format is:

- "title" to keep track of the section headings.

- "data" is the ingredient object from the pantry to make it easy to connect all of the different ingredients.

NOTE:
The initial grocery list is not like the figma with the plus sign on the left of the screen. Should be discussed if that is desired.